Add MYPLAYGROUND behavior to `CMakeLists.txt in faunus/src`
This particular piece of code was copied from `faunus/src/playground/CMakeLists.txt` to `faunus/src/CMakeLists.txt`

Since 388a5a1382 the `CMakeLists.txt` inside the `playground` folder is no longer executed. However I was relying on the behavior of this piece to code in a faunus-notebook, so I copied it to the `CMakeLists.txt` of the parent directory.